feat: 최단 경로 찾기 기능 구현

+++ src/main/java/wooteco/subway/domain/Fare.java
@@ -12,8 +12,8 @@ public class Fare {
 
     private final int fare;
 
-    public Fare(int distance) {
-        fare = calculateFare(distance);
+    public Fare(double distance) {
+        fare = calculateFare((int) distance);
     }
 
     private int calculateFare(int distance) {
